feat: Delete SubHeader
