+++ src/components/Header/HeaderLayout.module.scss
@@ -61,11 +61,15 @@
   li {
     float: left;
   }
-  span {
+  .gnb_element {
     color: white;
     font-size: 18px;
     font-weight: bold;
     line-height: 14px;
     padding: 15px 9px;
+    cursor: pointer;
+  }
+  a {
+    text-decoration: none;
   }
 }
