--- src/components/SubHeader/ChannelHeader.module.scss
@@ -1,33 +1,0 @@
-.channel {
-  align-items: center;
-  background-color: #f9f9f9;
-  border-bottom: 1px solid #ddd;
-  display: flex;
-  font-size: 14px;
-  height: 50px;
-  padding: 0 15px;
-  .channel_terrace {
-    padding: 5px;
-  }
-  .channel_aside {
-    display: flex;
-    margin-left: auto;
-  }
-  .channel_raffle {
-    align-items: center;
-    display: flex;
-    .channel_raffle_link {
-      color: #0078ff;
-      font-weight: 700;
-      margin-left: 10px;
-    }
-  }
-  .channel_raffle::after {
-    background-color: #d1d1d1;
-    content: '';
-    display: block;
-    height: 12px;
-    margin: 0 10px;
-    width: 1px;
-  }
-}
